% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{add_op_set_op}
\alias{add_op_set_op}
\title{Append a set operation to the tbl_kusto object's ops list}
\usage{
add_op_set_op(x, y, type)
}
\arguments{
\item{x}{The "left" tbl}

\item{y}{The "right" tbl}

\item{type}{The type of set operation to perform, currently only supports union_all}
}
\description{
Append a set operation to the tbl_kusto object's ops list
}
