% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{add_op_single}
\alias{add_op_single}
\title{Append an operation representing a single-table verb to the tbl_kusto object's ops list}
\usage{
add_op_single(name, .data, dots = list(), args = list())
}
\arguments{
\item{name}{The name of the operation, e.g. 'select', 'filter'}

\item{.data}{The tbl_kusto object to append the operation to}

\item{dots}{The expressions passed as arguments to the operation verb}

\item{args}{Other non-expression arguments passed to the operation verb}
}
\description{
Append an operation representing a single-table verb to the tbl_kusto object's ops list
}
