% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-kql.R
\docType{data}
\name{base_agg}
\alias{base_agg}
\title{Aggregation function translations}
\format{An object of class \code{environment} of length 7.}
\usage{
base_agg
}
\description{
Aggregation function translations
}
\keyword{datasets}
