% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-kql.R
\docType{data}
\name{base_scalar}
\alias{base_scalar}
\title{Scalar operator translations (infix and prefix)}
\format{
An object of class \code{environment} of length 75.
}
\usage{
base_scalar
}
\description{
Scalar operator translations (infix and prefix)
}
\keyword{datasets}
