% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-kql.R
\docType{data}
\name{base_window}
\alias{base_window}
\title{Window function translations}
\format{An object of class \code{environment} of length 1.}
\usage{
base_window
}
\description{
Window function translations
}
\keyword{datasets}
