% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kql-escape.R
\name{build_kql}
\alias{build_kql}
\title{Build a KQL string.}
\usage{
build_kql(..., .env = parent.frame(), con = NULL)
}
\arguments{
\item{...}{input to convert to KQL. Use \code{\link[=kql]{kql()}} to preserve
user input as is (dangerous), and \code{\link[=ident]{ident()}} to label user
input as kql identifiers (safe)}

\item{.env}{the environment in which to evalute the arguments. Should not
be needed in typical use.}

\item{con}{database connection; used to select correct quoting characters.}
}
\description{
Build a KQL string.
}
