% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{call_kusto}
\alias{call_kusto}
\title{Run a query or command against a Kusto database}
\usage{
call_kusto(database, qry_cmd, ..., .http_status_handler = "stop")
}
\arguments{
\item{database}{A Kusto database endpoint object, as returned by \code{kusto_database_endpoint}.}

\item{qry_cmd}{A string containing the query or command. In KQL, a database management command is a statement that starts with a "."}

\item{...}{Named arguments to be used as parameters for a parameterized query. These are ignored for database management commands.}
}
\description{
Run a query or command against a Kusto database
}
\details{
This function is the workhorse of the AzureKusto package. It communicates with the Kusto server and returns the query or command results, as data frames.
}
\seealso{
\link{kusto_database_endpoint}
}
