% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\docType{methods}
\name{dbSendQuery,AzureDataExplorerConnection-method}
\alias{dbSendQuery,AzureDataExplorerConnection-method}
\title{Send a query to AzureDataExplorer.}
\usage{
\S4method{dbSendQuery}{AzureDataExplorerConnection}(conn, statement, ...)
}
\description{
Send a query to AzureDataExplorer.
}
\examples{

}
