% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\docType{methods}
\name{dbSendQuery,AzureKustoConnection-method}
\alias{dbSendQuery,AzureKustoConnection-method}
\title{Send a query to AzureKusto.}
\usage{
\S4method{dbSendQuery}{AzureKustoConnection}(conn, statement, ...)
}
\description{
Send a query to AzureKusto.
}
\examples{

}
