% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{export.tbl_kusto}
\alias{export.tbl_kusto}
\title{Execute the Kusto query and export the result to Azure Storage.}
\usage{
\method{export}{tbl_kusto}(
  tbl,
  storage_uri,
  name_prefix = "export",
  key = "impersonate",
  format = "parquet",
  distributed = FALSE,
  ...
)
}
\arguments{
\item{tbl}{An instance of class tbl_kusto representing a Kusto table}

\item{storage_uri}{The Azure Storage URI to export files to.}

\item{name_prefix}{The filename prefix to use for exported files.}

\item{key}{default "impersonate" which uses the account signed into Kusto to
authenticate to Azure Storage. An Azure Storage account key.}

\item{format}{Options are "parquet", "csv", "tsv", "json"}

\item{distributed}{logical, indicates whether Kusto should distributed the
export job to multiple nodes, in which case multiple files will be written
to storage concurrently.}

\item{...}{needed for agreement with generic. Not otherwise used.}
}
\description{
Execute the Kusto query and export the result to Azure Storage.
}
