% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kql-build.R
\name{flatten_query}
\alias{flatten_query}
\title{Walks the tree of ops and builds a stack.}
\usage{
flatten_query(op, ops = list())
}
\arguments{
\item{op}{the current operation}

\item{ops}{the stack of operations to append to, recursively}
}
\description{
Walks the tree of ops and builds a stack.
}
