% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_ade_token}
\alias{get_ade_token}
\title{Obtain an AAD authentication token for the given Kusto cluster and tenant}
\usage{
get_ade_token(cluster, location = NULL, tenant, refresh = TRUE)
}
\description{
Obtain an AAD authentication token for the given Kusto cluster and tenant
}
