% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ident.R
\name{ident}
\alias{ident}
\title{Flag a character string as a Kusto identifier}
\usage{
ident(...)
}
\arguments{
\item{...}{character strings to flag as Kusto identifiers}
}
\description{
Flag a character string as a Kusto identifier
}
