% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ident.R
\name{ident_q}
\alias{ident_q}
\title{Pass an already-escaped string to Kusto}
\usage{
ident_q(...)
}
\arguments{
\item{...}{character strings to treat as already-escaped identifiers}
}
\description{
Pass an already-escaped string to Kusto
}
