% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_from_file}
\alias{ingest_from_file}
\alias{ingest_from_url}
\alias{ingest_from_blob}
\alias{ingest_from_adls2}
\title{Ingestion functions for Kusto}
\usage{
ingest_from_file(database, src, dest_table, streaming_ingest = FALSE,
  ...)

ingest_from_url(database, src, dest_table, async = FALSE, ...)

ingest_from_blob(database, src, key = NULL, token = NULL, sas = NULL,
  dest_table, async = FALSE, ...)

ingest_from_adls2(database, src, key = NULL, token = NULL,
  sas = NULL, dest_table, async = FALSE, ...)
}
\arguments{
\item{database}{A Kusto database endpoint object, created with \link{kusto_database_endpoint}.}

\item{src}{The source file or URL.}

\item{dest_table}{The name of the destination table.}

\item{streaming_ingest}{For file ingestion, whether to upload the file directly to the ingestion endpoint. If FALSE (the default), the file is instead uploaded to blob storage, and ingested from there.}

\item{...}{Named arguments to be treated as ingestion parameters.}

\item{async}{For the URL ingestion functions, whether to do the ingestion asychronously. If TRUE, the function will return immediately while the server handles the operation in the background.}

\item{key, token, sas}{Authentication arguments for the Azure storage ingestion methods. If multiple arguments are supplied, a key takes priority over a token, which takes priority over a SAS.}
}
\description{
Ingestion functions for Kusto
}
