% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_database.R, R/az_kusto.R
\name{is_kusto_database}
\alias{is_kusto_database}
\alias{is_kusto_cluster}
\title{Information functions}
\usage{
is_kusto_database(x)

is_kusto_cluster(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
These functions test whethen an object is of the given class.
}
