% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-kql.R
\name{kql}
\alias{kql}
\title{Tag character strings as Kusto Query Language. Assumes the string is valid and properly escaped.}
\usage{
kql(...)
}
\arguments{
\item{...}{character strings to tag as KQL}
}
\description{
Tag character strings as Kusto Query Language. Assumes the string is valid and properly escaped.
}
