% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-kql.R
\name{kql_aggregate}
\alias{kql_aggregate}
\title{Return a function representing a KQL aggregation function}
\usage{
kql_aggregate(f)
}
\arguments{
\item{f}{Name of the Kusto aggregation function}
}
\description{
Return a function representing a KQL aggregation function
}
