% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kql-build.R
\name{kql_build.op_mutate}
\alias{kql_build.op_mutate}
\title{dplyr's mutate verb can include aggregations, but Kusto's extend verb cannot.
If the mutate contains no aggregations, then it can emit an extend clause.
If the mutate contains an aggregation and the tbl is ungrouped,
then it must emit a summarize clause grouped by all variables.
If the mutate contains an aggregation and the tbl is grouped,
then it must join to a subquery containing the summarize clause.}
\usage{
\method{kql_build}{op_mutate}(op, ...)
}
\arguments{
\item{op}{A nested sequence of query operations, i.e. tbl_kusto$ops}

\item{...}{Needed for agreement with generic. Not otherwise used.}
}
\description{
dplyr's mutate verb can include aggregations, but Kusto's extend verb cannot.
If the mutate contains no aggregations, then it can emit an extend clause.
If the mutate contains an aggregation and the tbl is ungrouped,
then it must emit a summarize clause grouped by all variables.
If the mutate contains an aggregation and the tbl is grouped,
then it must join to a subquery containing the summarize clause.
}
