% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kql-build.R
\name{kql_build}
\alias{kql_build}
\title{Build the tbl object into a data structure representing a Kusto query}
\usage{
kql_build(op)
}
\arguments{
\item{op}{A nested sequence of query operations, i.e. tbl_kusto$ops}
}
\description{
Build the tbl object into a data structure representing a Kusto query
}
