% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kql-escape.R
\name{kql_escape_ident}
\alias{kql_escape_ident}
\title{Escape a Kusto identifer (pass-through, as identifiers don't need quoting)}
\usage{
kql_escape_ident(x)
}
\arguments{
\item{x}{An identifier to escape}
}
\description{
Escape a Kusto identifer (pass-through, as identifiers don't need quoting)
}
