% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kql-escape.R
\name{kql_escape_ident_q}
\alias{kql_escape_ident_q}
\title{Pass through an already-escaped Kusto identifier}
\usage{
kql_escape_ident_q(x)
}
\arguments{
\item{x}{An identifier to pass through}
}
\description{
Pass through an already-escaped Kusto identifier
}
