% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kql-escape.R
\name{kql_escape_logical}
\alias{kql_escape_logical}
\title{Escape a Kusto logical value.
Converts TRUE/FALSE to true / false}
\usage{
kql_escape_logical(x)
}
\arguments{
\item{x}{A logical value to escape}
}
\description{
Escape a Kusto logical value.
Converts TRUE/FALSE to true / false
}
