% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kql-escape.R
\name{kql_escape_string}
\alias{kql_escape_string}
\title{Escape a Kusto string by single-quoting}
\usage{
kql_escape_string(x)
}
\arguments{
\item{x}{A string to escape}
}
\description{
Escape a Kusto string by single-quoting
}
