% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-kql.R
\name{kql_infix}
\alias{kql_infix}
\title{Return a function representing a scalar KQL infix operator}
\usage{
kql_infix(f)
}
\arguments{
\item{f}{Name of a Kusto infix operator / function}
}
\description{
Return a function representing a scalar KQL infix operator
}
