% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-kql.R
\name{kql_prefix}
\alias{kql_prefix}
\title{Return a function representing a scalar KQL prefix function}
\usage{
kql_prefix(f, n = NULL)
}
\arguments{
\item{f}{Name of a Kusto infix function}

\item{n}{Number of arguments accepted by the Kusto prefix function}
}
\description{
Return a function representing a scalar KQL prefix function
}
