% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kql-escape.R
\name{kql_quote}
\alias{kql_quote}
\title{Helper function for quoting kql elements.}
\usage{
kql_quote(x, quote)
}
\arguments{
\item{x}{Character vector to escape.}

\item{quote}{Single quoting character.}
}
\description{
If the quote character is present in the string, it will be doubled.
\code{NA}s will be replaced with NULL.
}
\examples{
kql_quote("abc", "'")
kql_quote("I've had a good day", "'")
kql_quote(c("abc", NA), "'")
}
\keyword{internal}
