% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kql-render.R
\name{kql_render}
\alias{kql_render}
\title{Render a set of operations on a tbl_kusto_abstract to a Kusto query}
\usage{
kql_render(query, ...)
}
\arguments{
\item{query}{The tbl_kusto instance with a sequence of operations in $ops}

\item{...}{needed for agreement with generic. Not otherwise used.}
}
\description{
Render a set of operations on a tbl_kusto_abstract to a Kusto query
}
