% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kql-render.R
\name{kql_render.tbl_kusto_abstract}
\alias{kql_render.tbl_kusto_abstract}
\title{Render a set of operations on a tbl_kusto_abstract to a Kusto query}
\usage{
\method{kql_render}{tbl_kusto_abstract}(query, ...)
}
\arguments{
\item{query}{The tbl_kusto instance with a sequence of operations in $ops}
}
\description{
Render a set of operations on a tbl_kusto_abstract to a Kusto query
}
