% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-kql.R
\name{kql_translate_env}
\alias{kql_translate_env}
\title{Build a kql_variant class out of the environments holding scalar and aggregation
function definitions}
\usage{
kql_translate_env()
}
\description{
Build a kql_variant class out of the environments holding scalar and aggregation
function definitions
}
