% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-kql.R
\name{kql_translator}
\alias{kql_translator}
\title{Builds an environment from a list of R -> Kusto query language translation pairs.}
\usage{
kql_translator(..., .funs = list(), .parent = new.env(parent = emptyenv()))
}
\arguments{
\item{...}{Pairs of R call = Kusto call translations as individual arguments}

\item{.funs}{Parse of R call = Kusto call translations in list format}

\item{.parent}{A parent environment to attach this env onto}
}
\description{
Builds an environment from a list of R -> Kusto query language translation pairs.
}
