% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-kql.R
\name{kql_window}
\alias{kql_window}
\title{Return a function representing a KQL window function}
\usage{
kql_window(f)
}
\arguments{
\item{f}{Name of the Kusto aggregation function}
}
\description{
Return a function representing a KQL window function
}
