% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{kusto_export_cmd}
\alias{kusto_export_cmd}
\title{Run a Kusto query and export results to Azure Storage in Parquet or CSV
format.}
\usage{
kusto_export_cmd(query, storage_uri, name_prefix, key, format, distributed)
}
\arguments{
\item{query}{The text of the Kusto query to run}

\item{storage_uri}{The URI of the blob storage container to export to}

\item{name_prefix}{The filename prefix for each exported file}

\item{key}{The account key for the storage container.
uses the identity that is signed into Kusto to authenticate to Azure Storage.}

\item{format}{Options are "parquet", "csv", "tsv", "json"}

\item{distributed}{logical, indicates whether Kusto should distributed the
export job to multiple nodes, in which case multiple files will be written
to storage concurrently.}
}
\description{
Run a Kusto query and export results to Azure Storage in Parquet or CSV
format.
}
