% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{nest.tbl_kusto_abstract}
\alias{nest.tbl_kusto_abstract}
\title{Nest method for Kusto tables}
\usage{
nest.tbl_kusto_abstract(.data, ...)
}
\arguments{
\item{.data}{A kusto tbl.}

\item{...}{Specification of columns to nest. Translates to summarize make_list() in Kusto.}
}
\description{
This method collapses a column into a list
}
