% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_double}
\alias{op_double}
\title{A double-table verb, e.g. joins, setops}
\usage{
op_double(name, x, y, args = list())
}
\arguments{
\item{name}{The name of the operation, e.g. 'left_join', 'union_all'}

\item{x}{The "left" tbl}

\item{y}{The "right" tbl}

\item{args}{Other arguments passed to the operation verb}
}
\description{
A double-table verb, e.g. joins, setops
}
