% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_grps}
\alias{op_grps}
\title{Look up the applicable grouping variables for an operation
based on the data source and preceding sequence of operations}
\usage{
op_grps(op)
}
\arguments{
\item{op}{An operation instance}
}
\description{
Look up the applicable grouping variables for an operation
based on the data source and preceding sequence of operations
}
