% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_single}
\alias{op_single}
\title{A class representing a single-table verb}
\usage{
op_single(name, x, dots = list(), args = list())
}
\arguments{
\item{name}{the name of the operation verb, e.g. "select", "filter"}

\item{x}{the tbl object}

\item{dots}{expressions passed to the operation verb function}

\item{args}{other arguments passed to the operation verb function}
}
\description{
A class representing a single-table verb
}
