% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_vars}
\alias{op_vars}
\title{Look up the applicable variables in scope for a given operation
based on the data source and preceding sequence of operations}
\usage{
op_vars(op)
}
\arguments{
\item{op}{An operation instance}
}
\description{
Look up the applicable variables in scope for a given operation
based on the data source and preceding sequence of operations
}
