% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{run_query}
\alias{run_query}
\alias{run_query.kusto_database_endpoint}
\alias{run_command}
\alias{run_command.kusto_database_endpoint}
\title{Run a query or command against a Kusto database}
\usage{
run_query(database, ...)

\method{run_query}{kusto_database_endpoint}(database, query, ...)

run_command(database, ...)

\method{run_command}{kusto_database_endpoint}(database, command, ...)
}
\arguments{
\item{database}{A Kusto database endpoint object, as returned by \code{kusto_query_endpoint}.}

\item{...}{For \code{run_query}, named arguments to be used as parameters for a parameterized query.}

\item{query, command}{A string containing the query or command. Note that database management commands in KQL are distinct from queries.}
}
\description{
Run a query or command against a Kusto database
}
\details{
These functions are the workhorses of the AzureKusto package. They communicate with the Kusto server and return the query or command results, as data frames.
}
\seealso{
\link{kusto_query_endpoint}
}
