% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{simulate_kusto}
\alias{simulate_kusto}
\title{Simulate a kusto database endpoint as the source for a tbl_kusto_abstract.}
\usage{
simulate_kusto()
}
\description{
Simulate a kusto database endpoint as the source for a tbl_kusto_abstract.
}
