% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{tbl.tbl_df}
\alias{tbl.tbl_df}
\title{Create a dummy tbl from a data frame.
Useful for testing KQL generation without a remote connection.}
\usage{
\method{tbl}{tbl_df}(table, name, object = simulate_ade())
}
\description{
Create a dummy tbl from a data frame.
Useful for testing KQL generation without a remote connection.
}
