% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{tbl_abstract}
\alias{tbl_abstract}
\title{Create a local lazy tbl}
\usage{
tbl_abstract(df, src = NULL)
}
\description{
Useful for testing KQL generation without a remote connection.
}
\examples{
library(dplyr)
df <- data.frame(x = 1, y = 2)

df <- tbl_lazy(df, src = simulate_ade())
df \%>\% summarise(x = sd(x)) \%>\% show_query()
}
\keyword{internal}
