% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{tbl_kusto}
\alias{tbl_kusto}
\title{A tbl object representing a table in a Kusto database.}
\usage{
tbl_kusto(kusto_database, table_name, ...)
}
\arguments{
\item{kusto_database}{An instance of kusto_database_endpoint that this table should be queried from}

\item{table_name}{The name of the table in the Kusto database}

\item{...}{needed for agreement with generic. Not otherwise used.}
}
\description{
A tbl object representing a table in a Kusto database.
}
