% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{tbl_kusto_abstract}
\alias{tbl_kusto_abstract}
\title{Create a local lazy tbl}
\usage{
tbl_kusto_abstract(df, table_name, ...)
}
\description{
Useful for testing KQL generation without a remote connection.
}
\examples{
library(dplyr)
df <- data.frame(x = 1, y = 2)

df <- tbl_kusto_abstract(df, "table1")
df \%>\%
    summarise(x = sd(x)) \%>\%
    show_query()
}
\keyword{internal}
