% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-kql.R
\name{translate_kql}
\alias{translate_kql}
\title{Translate R expressions into Kusto Query Language equivalents.}
\usage{
translate_kql(...)
}
\arguments{
\item{...}{Expressions to translate.}
}
\description{
Translate R expressions into Kusto Query Language equivalents.
}
