% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{add_op_join}
\alias{add_op_join}
\title{Append a join operation to the tbl_kusto object's ops list}
\usage{
add_op_join(type, x, y, by = NULL, suffix = NULL)
}
\arguments{
\item{type}{The name of the join type,
one of: inner_join, left_join, right_join, full_join, semi_join, anti_join}

\item{x}{The "left" tbl}

\item{y}{The "right" tbl}

\item{by}{A vector of column names; keys by which tbl x and tbl y will be joined}

\item{suffix}{A vector of strings that will be appended to the names of non-join key columns that exist in both tbl x and tbl y to distinguish them by source tbl.}
}
\description{
Append a join operation to the tbl_kusto object's ops list
}
