% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\docType{methods}
\name{dbConnect,AzureKustoDriver-method}
\alias{dbConnect,AzureKustoDriver-method}
\title{Connect to a Kusto cluster}
\usage{
\S4method{dbConnect}{AzureKustoDriver}(drv, ...)
}
\description{
Connect to a Kusto cluster
}
\examples{
\dontrun{
db <- dbConnect(AzureKusto::AzureKusto(),
                server="https://mycluster.location.kusto.windows.net", database="database"...)
dbWriteTable(db, "mtcars", mtcars)
dbGetQuery(db, "mtcars | where cyl == 4")
}
}
