% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{collect.tbl_kusto}
\alias{collect.tbl_kusto}
\title{Compile the preceding dplyr operations into a kusto query, execute it on the remote server,
and return the result as a tibble.}
\usage{
\method{collect}{tbl_kusto}(x, ...)
}
\arguments{
\item{x}{An instance of class tbl_kusto representing a Kusto table}

\item{...}{needed for agreement with generic. Not otherwise used.}
}
\description{
Compile the preceding dplyr operations into a kusto query, execute it on the remote server,
and return the result as a tibble.
}
