% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kql-escape.R
\name{escape}
\alias{escape}
\alias{kql_vector}
\title{Escape/quote a string.}
\usage{
escape(x, parens = NA, collapse = " ")

kql_vector(x, parens = NA, collapse = " ")
}
\arguments{
\item{x}{An object to escape. Existing kql vectors will be left as is,
character vectors are escaped with single quotes, numeric vectors have
trailing \code{.0} added if they're whole numbers, identifiers are
escaped with double quotes.}

\item{parens, collapse}{Controls behaviour when multiple values are supplied.
\code{parens} should be a logical flag, or if \code{NA}, will wrap in
parens if length > 1.

Default behaviour: lists are always wrapped in parens and separated by
commas, identifiers are separated by commas and never wrapped,
atomic vectors are separated by spaces and wrapped in parens if needed.}
}
\description{
Escape/quote a string.
}
