% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_kusto_token}
\alias{get_kusto_token}
\title{Obtain an AAD authentication token for the given Kusto cluster and tenant}
\usage{
get_kusto_token(cluster, location = NULL, tenant, refresh = TRUE)
}
\description{
Obtain an AAD authentication token for the given Kusto cluster and tenant
}
