% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_base}
\alias{op_base}
\title{The "base case" operation representing the tbl itself and its column variables}
\usage{
op_base(x, vars, class = character())
}
\arguments{
\item{x}{A tbl object}

\item{vars}{A vector of column variables in the tbl}
}
\description{
The "base case" operation representing the tbl itself and its column variables
}
