% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-dplyr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{intersect}
\alias{setdiff}
\alias{setequal}
\alias{union}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr:setops]{intersect}}, \code{\link[dplyr:setops]{setdiff}}, \code{\link[dplyr:setops]{setequal}}, \code{\link[dplyr:setops]{union}}}
}}

