% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{tbl_kusto}
\alias{tbl_kusto}
\title{A tbl object representing a table in a Kusto database.}
\usage{
tbl_kusto(kusto_database, table_name, ...)
}
\arguments{
\item{...}{needed for agreement with generic. Not otherwise used.}

\item{subclass}{name of subclass}
}
\description{
A tbl object representing a table in a Kusto database.
}
