% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{unnest.tbl_kusto_abstract}
\alias{unnest.tbl_kusto_abstract}
\title{Unnest method for Kusto tables}
\usage{
unnest.tbl_kusto_abstract(.data, ..., .id = NULL)
}
\arguments{
\item{.data}{A Kusto tbl.}

\item{...}{Specification of columns to unnest.}

\item{.id}{Data frame identifier - if supplied, will create a new column with name .id, giving a unique identifier. This is most useful if the list column is named.}
}
\description{
This method takes a list column and expands it so that each element of the list gets its own row.
unnest() translates to Kusto's mv-expand operator.
}
